+++ backend/repositories/meanings.py
@@ -6,9 +6,10 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from backend.models import Meaning
 from backend.repositories import Repository
-from backend.schemas import MeaningCreate, MeaningUpdate, ParamsMeaning
+from backend.schemas import MeaningCreate, MeaningUpdate, ParamsMeaning, PermissionType, UserAuth
 
 from .words import Words
+
 
 # https://stackoverflow.com/questions/68360687/sqlalchemy-asyncio-orm-how-to-query-the-database
 
@@ -39,15 +40,17 @@ class Meanings(Repository):
 
         return meanings
 
-    async def create(self, entity: MeaningCreate) -> None:
+    async def create(self, word_id: int, entity: MeaningCreate, user: UserAuth) -> None:
         meaning_db = Meaning(
             meaning=entity.meaning,
             comment=entity.comment,
             chapter_id=entity.chapter_id,
             entry_id=entity.entry_id,
+            user_id=user.id,
             word_id=entity.word_id,
             reference_id=entity.reference_id,
         )
+
         self.session.add(meaning_db)
         await self.session.commit()
         await self.session.refresh(meaning_db)
@@ -64,7 +67,7 @@ class Meanings(Repository):
         return meaning
 
     async def update_by_id(
-        self, entity_id: int, entity: MeaningUpdate
+            self, entity_id: int, entity: MeaningUpdate
     ) -> None:
         meaning_db = await self.get_by_id(entity_id)
 
@@ -78,14 +81,23 @@ class Meanings(Repository):
         await self.session.commit()
         await self.session.refresh(meaning_db)
 
-    async def delete_by_id(self, entity_id: int) -> None:
+    async def delete_by_id(self, entity_id: int, user: UserAuth) -> None:
         meaning_db = await self.get_by_id(entity_id)
+
+        if (
+                meaning_db.user_id != user.id
+                and user.permission != PermissionType.ADMIN
+        ):
+            raise HTTPException(
+                status_code=status.HTTP_403_FORBIDDEN,
+                detail='Usuário sem permissão.',
+            )
 
         await self.session.delete(meaning_db)
         await self.session.commit()
 
     async def get_list_by_word_id(
-        self, word_id: int, params: ParamsMeaning
+            self, word_id: int, params: ParamsMeaning
     ) -> Sequence[Meaning]:
         await self.words.get_by_id(word_id)
 
@@ -111,17 +123,8 @@ class Meanings(Repository):
 
         return meanings
 
-    async def create_by_word(
-        self, word_id: int, entity: MeaningCreate
-    ) -> None:
-        meaning_db = Meaning(
-            meaning=entity.meaning,
-            comment=entity.comment,
-            chapter_id=entity.chapter_id,
-            entry_id=entity.entry_id,
-            word_id=word_id,
-            reference_id=entity.reference_id,
-        )
-        self.session.add(meaning_db)
-        await self.session.commit()
-        await self.session.refresh(meaning_db)
+    async def all(self) -> Sequence[Meaning]:
+        statement = select(Meaning)
+        result = await self.session.execute(statement)
+        meanings = result.scalars().all()
+        return meanings
